+++ script/turtle.py
@@ -9,7 +9,7 @@ class Turtle:
     
     # move forward
     def forward(self, dist: np.float64) -> None:
-        self.pos += dist * np.array((math.cos(math.radians(self.heading)), math.sin(math.radians(self.heading))), dtype=np.float16)
+        self.pos += dist * np.array((math.cos(math.radians(self.heading)), math.sin(math.radians(self.heading))), dtype=np.float64)
 
     # turn right
     def right(self, angle: np.float64) -> None:
